update to fuelphp version 1.72

+++ fuel/core/classes/ftp.php
@@ -3,10 +3,10 @@
  * Part of the Fuel framework.
  *
  * @package    Fuel
- * @version    1.5
+ * @version    1.7
  * @author     Fuel Development Team
  * @license    MIT License
- * @copyright  2010 - 2013 Fuel Development Team
+ * @copyright  2010 - 2014 Fuel Development Team
  * @link       http://fuelphp.com
  */
 
@@ -272,7 +272,7 @@ class Ftp
 			return false;
 		}
 
-		if ( ! file_exists($local_path))
+		if ( ! is_file($local_path))
 		{
 			throw new \FtpFileAccessException('No source file');
 			return false;
@@ -459,7 +459,11 @@ class Ftp
 				// we'll recursively call delete_dir()
 				if ( ! @ftp_delete($this->_conn_id, $item))
 				{
-					$this->delete_dir($item);
+					// don't recurse into current of parent directory
+					if ( ! preg_match('/\/\.\.|\/\.$/', $item)) 
+					{
+						$this->delete_dir($item);
+					}
 				}
 			}
 		}
